realstatus/requirements.txt
django
djangorestframework
djangorestframework-jwt
djangorestframework-simplejwt
django-ninja
gunicorn
fastapi

typing-extensions==3.10.0.0
django-debug-toolbar==3.2.1

numpy
pandas 
tensorflow
keras
mysqlclient